requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements-dev.in
#
astor==0.8.1
    # via flake8-qgis
certifi==2024.7.4
    # via requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
    # via pytest
coverage[toml]==7.5.1
    # via pytest-cov
debugpy==1.8.1
    # via -r requirements-dev.in
distlib==0.3.8
    # via virtualenv
filelock==3.14.0
    # via virtualenv
flake8==7.0.0
    # via
    #   -r requirements-dev.in
    #   flake8-qgis
flake8-qgis==1.0.0
    # via -r requirements-dev.in
identify==2.5.36
    # via pre-commit
idna==3.7
    # via requests
importlib-metadata==8.2.0
    # via qgis-plugin-dev-tools
iniconfig==2.0.0
    # via pytest
mccabe==0.7.0
    # via flake8
mypy==1.10.0
    # via -r requirements-dev.in
mypy-extensions==1.0.0
    # via mypy
nodeenv==1.8.0
    # via pre-commit
packaging==24.0
    # via
    #   pytest
    #   qgis-plugin-dev-tools
platformdirs==4.2.1
    # via virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==3.7.1
    # via -r requirements-dev.in
pycodestyle==2.11.1
    # via flake8
pyflakes==3.2.0
    # via flake8
pyqt5-stubs==5.15.6.0
    # via -r requirements-dev.in
pytest==8.2.0
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-qgis
pytest-cov==5.0.0
    # via -r requirements-dev.in
pytest-qgis==2.0.0
    # via -r requirements-dev.in
python-dotenv==1.0.1
    # via qgis-plugin-dev-tools
pyyaml==6.0.1
    # via pre-commit
qgis-plugin-dev-tools==0.7.0
    # via -r requirements-dev.in
requests==2.32.3
    # via qgis-plugin-dev-tools
ruff==0.4.4
    # via -r requirements-dev.in
tomli==2.0.1
    # via qgis-plugin-dev-tools
typing-extensions==4.11.0
    # via mypy
urllib3==2.2.2
    # via requests
virtualenv==20.26.2
    # via pre-commit
zipp==3.19.2
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
